requirements.txt
Flask>=3.0,<4.0
python-chess>=1.999
gunicorn>=21.0,<24.0
